# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "家庭故事豆 API",
    "description": "家庭故事豆 MVP 的 API 契约与 Mock 联调服务。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/podcasts": {
      "post": {
        "tags": [
          "podcasts"
        ],
        "summary": "Create Podcast",
        "operationId": "create_podcast_v1_podcasts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePodcastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/podcasts/{podcast_id}": {
      "get": {
        "tags": [
          "podcasts"
        ],
        "summary": "Get Podcast",
        "operationId": "get_podcast_v1_podcasts__podcast_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "podcast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Podcast Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "podcasts"
        ],
        "summary": "Remix Podcast",
        "operationId": "remix_podcast_v1_podcasts__podcast_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "podcast_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Podcast Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemixPodcastRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/recordings/{recording_id}/podcast-shares": {
      "post": {
        "tags": [
          "podcast-sharing"
        ],
        "summary": "Create Podcast Share",
        "operationId": "create_podcast_share_v1_recordings__recording_id__podcast_shares_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Recording Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePodcastShareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastShareData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/podcast-shares": {
      "get": {
        "tags": [
          "podcast-sharing"
        ],
        "summary": "List Podcast Shares",
        "operationId": "list_podcast_shares_v1_podcast_shares_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "recording_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Recording Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_PodcastShareData__"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/podcast-shares/{share_id}": {
      "delete": {
        "tags": [
          "podcast-sharing"
        ],
        "summary": "Revoke Podcast Share",
        "operationId": "revoke_podcast_share_v1_podcast_shares__share_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PodcastShareData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/public/podcast-shares/{token}": {
      "get": {
        "tags": [
          "podcast-sharing"
        ],
        "summary": "Resolve Podcast Share",
        "operationId": "resolve_podcast_share_v1_public_podcast_shares__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PublicPodcastShareData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/public/podcast-shares/{token}/card": {
      "get": {
        "tags": [
          "podcast-sharing"
        ],
        "summary": "Podcast Share Card",
        "operationId": "podcast_share_card_v1_public_podcast_shares__token__card_get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/creations/{creation_id}/shares": {
      "post": {
        "tags": [
          "timeline-privacy"
        ],
        "summary": "Create Share",
        "operationId": "create_share_v1_creations__creation_id__shares_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "creation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Creation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateShareRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ShareData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/shares": {
      "get": {
        "tags": [
          "timeline-privacy"
        ],
        "summary": "List Shares",
        "operationId": "list_shares_v1_shares_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_list_ShareData__"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/shares/{share_id}": {
      "delete": {
        "tags": [
          "timeline-privacy"
        ],
        "summary": "Revoke Share",
        "operationId": "revoke_share_v1_shares__share_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "share_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Share Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_ShareData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/public/shares/{token}": {
      "get": {
        "tags": [
          "timeline-privacy"
        ],
        "summary": "Public Share",
        "operationId": "public_share_v1_public_shares__token__get",
        "parameters": [
          {
            "name": "token",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_PublicShareData_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiMeta": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "timestamp"
        ],
        "title": "ApiMeta"
      },
...
      "ApiResponse_PodcastData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/PodcastData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[PodcastData]"
      },
...
      "ApiResponse_PodcastShareData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/PodcastShareData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[PodcastShareData]"
      },
...
      "ApiResponse_PublicPodcastShareData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/PublicPodcastShareData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[PublicPodcastShareData]"
      },
      "ApiResponse_PublicShareData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/PublicShareData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[PublicShareData]"
      },
...
      "ApiResponse_ShareData_": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "$ref": "#/components/schemas/ShareData"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[ShareData]"
      },
...
      "ApiResponse_list_PodcastShareData__": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/PodcastShareData"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[PodcastShareData]]"
      },
...
      "ApiResponse_list_ShareData__": {
        "properties": {
          "ok": {
            "type": "boolean",
            "const": true,
            "title": "Ok",
            "default": true
          },
          "data": {
            "items": {
              "$ref": "#/components/schemas/ShareData"
            },
            "type": "array",
            "title": "Data"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiMeta"
          }
        },
        "type": "object",
        "required": [
          "data",
          "meta"
        ],
        "title": "ApiResponse[list[ShareData]]"
      },
...
      "CreatePodcastRequest": {
        "properties": {
          "moment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Moment Ids"
          }
        },
        "type": "object",
        "required": [
          "moment_ids"
        ],
        "title": "CreatePodcastRequest"
      },
      "CreatePodcastShareRequest": {
        "properties": {
          "expires_in_hours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Expires In Hours",
            "default": 24
          }
        },
        "type": "object",
        "title": "CreatePodcastShareRequest"
      },
...
      "CreateShareRequest": {
        "properties": {
          "expires_in_hours": {
            "type": "integer",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Expires In Hours",
            "default": 24
          }
        },
        "type": "object",
        "title": "CreateShareRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PodcastData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "pipeline_version": {
            "type": "string",
            "title": "Pipeline Version"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "moments": {
            "items": {
              "$ref": "#/components/schemas/PodcastMomentData"
            },
            "type": "array",
            "title": "Moments"
          },
          "segments": {
            "items": {
              "$ref": "#/components/schemas/PodcastSegmentData"
            },
            "type": "array",
            "title": "Segments"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "recording_id",
          "title",
          "status",
          "version",
          "pipeline_version",
          "metadata",
          "moments",
          "segments",
          "created_at",
          "updated_at"
        ],
        "title": "PodcastData"
      },
...
      "PodcastShareData": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "podcast_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Podcast Version Id"
          },
          "recording_id": {
            "type": "string",
            "format": "uuid",
            "title": "Recording Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "access_count": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Access Count"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "podcast_version_id",
          "recording_id",
          "title",
          "url",
          "expires_at",
          "access_count",
          "created_at"
        ],
        "title": "PodcastShareData"
      },
...
      "PublicPodcastShareData": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "title": "Tags"
          },
          "cover_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Url"
          },
          "cover_download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Download Url"
          },
          "media_url": {
            "type": "string",
            "title": "Media Url"
          },
          "expires_at": {
            "type": "string",
            "format": "date-time",
            "title": "Expires At"
          },
          "duration_ms": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Duration Ms"
          },
          "render_mode": {
            "type": "string",
            "enum": [
              "narrated",
              "original_only"
            ],
            "title": "Render Mode"
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "media_url",
          "expires_at",
          "duration_ms",
          "render_mode"
        ],
        "title": "PublicPodcastShareData"
      },
...
      "RemixPodcastRequest": {
        "properties": {
          "moment_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "maxItems": 3,
            "minItems": 1,
            "title": "Moment Ids"
          },
          "intro": {
            "type": "string",
            "maxLength": 600,
            "minLength": 1,
            "title": "Intro"
          },
          "outro": {
            "type": "string",
            "maxLength": 600,
            "minLength": 1,
            "title": "Outro"
          }
        },
        "type": "object",
        "required": [
          "moment_ids",
          "intro",
          "outro"
        ],
        "title": "RemixPodcastRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}